Fixed bug with Video and Pictures folder

+++ timelapse.py
@@ -3,6 +3,12 @@ from os import system
 import datetime
 from time import sleep
 import sys
+import os
+
+
+home_folder = os.getenv('HOME')
+files = os.listdir(home_folder)
+Pictures = os.listdir("{}/Pictures".format(home_folder))
 
 make_video = input("would you like to compile the video? (yes/no) -- If you are using a piZero (no) is recommended: ")
 
@@ -41,10 +47,19 @@ dateraw= datetime.datetime.now()
 datetimeformat = dateraw.strftime("%Y-%m-%d_%H:%M")
 print("RPi started taking photos for your timelapse at: " + datetimeformat)
 
+
+if not "Pictures" in files:
+    try:
+        os.mkdir("{}/Pictures".format(home_folder))
+        print("Created Pictures folder in User Home Directory")
+    except OSError as err:
+        print(err)
+
 camera = PiCamera()
 camera.resolution = (1920, 1080)
 
-system('rm /home/pi/Pictures/*.jpg') #delete all photos in the Pictures folder before timelapse start
+if Pictures:
+    system('rm /home/pi/Pictures/*.jpg') #delete all photos in the Pictures folder before timelapse start
 
 for i in range(int(numphotos)):
     camera.capture('/home/pi/Pictures/image{0:06d}.jpg'.format(i))
@@ -52,6 +67,14 @@ for i in range(int(numphotos)):
 print("Done taking photos.")
 
 if make_video == "yes":
+
+    if not "Videos" in files:
+        try:
+            os.mkdir("{}/Videos".format(home_folder))
+            print("Created Videos folder in User Home Directory")
+        except OSError as err:
+            print(err)
+
     print("Please standby as your timelapse video is created.")
     system('ffmpeg -r {} -f image2 -s 1024x768 -nostats -loglevel 0 -pattern_type glob -i "/home/pi/Pictures/*.jpg" -vcodec libx264 -crf 25  -pix_fmt yuv420p /home/pi/Videos/{}.mp4'.format(fps, datetimeformat))
     print('Timelapse video is complete. Video saved as /home/pi/Videos/{}.mp4'.format(datetimeformat))
